requirements.txt
Cython==0.29.23
apache-airflow-providers-google==4.0.0
apache-airflow-providers-http==2.0.0
beautifulsoup4==4.9.3
black==20.8b1
certifi==2021.5.30
gevent==21.8.0
google-cloud-firestore==2.1.3
lxml==4.6.3
pandas==1.2.4
pylint==2.9.6
pytest==6.2.3
pytest-cov==2.12.1
requests==2.25.1
types-requests==2.25.1
